feat: Fix assign reviewers
Previously fetch all applications from db which is inefficient. In this PR, i
utilize the permission field in user to mark if an application has been assigned
to be in_review. This way, it is more efficient since we are fetching only the
application that has permission of ['hacker'] which indicates that the
application has yet to be assigned. All assigned applications have permission of
['hacker','in_review'].